Fix BROOKLYN-493 Rebind fails when using WinRmCommandSensor
Add tests for CmdFeed rebind
